organize react components
