+++ src/react/design/animations/index.ts
@@ -1,0 +1,2 @@
+export * from './fade-in-up';
+export * from './slide-up';
