+++ src/react/design/index.ts
@@ -1,1 +1,2 @@
+export * from './animations';
 export * from './theme';
