+++ src/react/ui/index.ts
@@ -1,5 +1,7 @@
 export * from './Button';
 export * from './Form';
+export * from './ImageCrossfader';
 export * from './Input';
 export * from './LoadingSpinner';
+export * from './Modal';
 export * from './Navbar';
